up to find_objs (although I should take a pass at simplifying global_skysub)
